+++ swing/framework-server-monitor/src/main/java/is/codion/swing/framework/server/monitor/ui/ClientMonitorPanel.java
@@ -142,7 +142,7 @@ public final class ClientMonitorPanel extends JPanel {
 										.name("Disconnect")
 										.enabled(model.clientInstanceTableModel().selectionModel().selectionNotEmpty()))
 						.separator()
-						.controls(Controls.builder()
+						.control(Controls.builder()
 										.name("Columns")
 										.control(table.createToggleColumnsControls())
 										.control(table.createResetColumnsControl())